moved read/write to a new Codec struct and cloned stream for Tx / Rx
The solution was to clone the stream instance and use one for writing
and the other for reading. This should still point to the same stream
but allow simultaneous use.

This makes reading and writing packets agnostic to the protocol. The
handle_request function only needs to care about the content of packets
and not how they are wrapped.

Also, the client caching might have to be deprecated. It was made under
the assumption that data needs to be kept across multiple connections,
as if the ping request (for instance) was done in a different connection
than the status request. However this seems to be false.

Signed-off-by: kalmenn <kalmenn@proton.me>